+++ example/lib/main.dart
@@ -42,6 +42,8 @@ class _ExampleLauncherPageState extends State<ExampleLauncherPage> {
   String paymentMethodsPageResultText = '';
   String driversLicenseResultText = '';
   String photoUploadResultText = '';
+  String studentStatusResultText = '';
+  String verificationsResultText = '';
 
   BehaviorSubject<String?> userName = BehaviorSubject.seeded(null);
   BehaviorSubject<String?> userId = BehaviorSubject.seeded(null);
@@ -125,6 +127,16 @@ class _ExampleLauncherPageState extends State<ExampleLauncherPage> {
                   child: const Text('Payment Methods Page'),
                 ),
                 Text(paymentMethodsPageResultText),
+                ElevatedButton(
+                  onPressed: _openStudentStatusPage,
+                  child: const Text('Student Status'),
+                ),
+                Text(studentStatusResultText),
+                ElevatedButton(
+                  onPressed: _openVerificationsPage,
+                  child: const Text('Verifications'),
+                ),
+                Text(verificationsResultText),
                 Padding(
                   padding: const EdgeInsets.all(8.0),
                   child: Column(
@@ -260,6 +272,33 @@ class _ExampleLauncherPageState extends State<ExampleLauncherPage> {
     );
     setState(() {
       photoUploadResultText = 'Photo upload was shown, result: $result';
+    });
+  }
+
+  Future<void> _openStudentStatusPage() async {
+    setState(() {
+      studentStatusResultText = 'Opening Student status page...';
+    });
+
+    final result = await mLogin.openStudentStatus(
+      ephemeral: ephemeral,
+    );
+    setState(() {
+      studentStatusResultText = 'Student status was shown, result: $result';
+    });
+  }
+
+  Future<void> _openVerificationsPage() async {
+    setState(() {
+      verificationsResultText = 'Opening verifications page...';
+    });
+
+    final result = await mLogin.openVerifications(
+      ephemeral: ephemeral,
+      verificationType: const PhotoVerification(),
+    );
+    setState(() {
+      verificationsResultText = 'Verifications page was shown, result: $result';
     });
   }
 }