+++ lib/m_login_sdk.dart
@@ -5,6 +5,7 @@ export 'src/m_login_config.dart';
 export 'src/m_login_log.dart';
 export 'src/m_login_result.dart';
 export 'src/m_login_cit_result.dart';
+export 'src/m_login_verification_type.dart';
 export 'src/ui_elements/m_login_button.dart';
 export 'src/ui_elements/m_login_button_data.dart';
 export 'src/ui_elements/m_login_button_style.dart';