+++ lib/src/m_login.dart
@@ -8,6 +8,7 @@ import 'package:m_login_sdk/src/m_login_cit_result.dart';
 import 'package:m_login_sdk/src/m_login_config.dart';
 import 'package:m_login_sdk/src/m_login_log.dart';
 import 'package:m_login_sdk/src/m_login_result.dart';
+import 'package:m_login_sdk/src/m_login_verification_type.dart';
 
 ///
 /// Central access point to the MLoginSDK. All functionality is triggered via
@@ -242,6 +243,49 @@ class MLogin {
       ephemeral: ephemeral,
       username: prefilledUsername,
       overrideRedirectUri: overrideRedirectUri,
+    );
+  }
+
+  /// This function is used to authorize users for offers which are only
+  /// eligible for students, e.g. discounted tickets.
+  /// Shows a list of participating universities and colleges, which are
+  /// internally using the Shibboleth authentication.
+  /// User selects a university or college and is forwarded to its login.
+  ///
+  /// After successfully logging in, returns [true] and the user is
+  /// validated as a student. Returns [false] otherwise.
+  Future<bool> openStudentStatus({
+    bool ephemeral = false,
+    String? overrideRedirectUri,
+  }) {
+    return openDataPage(
+      this,
+      portalUriSuffix: 'verification/student/start',
+      ephemeral: ephemeral,
+      username: prefilledUsername,
+      overrideRedirectUri: overrideRedirectUri,
+    );
+  }
+
+  /// This function is used to display all verifications the user has completed
+  /// in her M-Login account.
+  /// By supplying a [MLoginVerificationType], the user can be guided to the
+  /// anchor in the page where this specific verification is shown.
+  ///
+  /// After successfully logging in, returns [true] and the user is
+  /// validated as a student. Returns [false] otherwise.
+  Future<bool> openVerifications({
+    bool ephemeral = false,
+    String? overrideRedirectUri,
+    MLoginVerificationType? verificationType,
+  }) {
+    return openDataPage(
+      this,
+      portalUriSuffix: 'verifications',
+      ephemeral: ephemeral,
+      username: prefilledUsername,
+      overrideRedirectUri: overrideRedirectUri,
+      anchor: verificationType?.anchor,
     );
   }